# Optimize GCLDAModel with numba