<!---
This is a suggested pull request template for NiMARE.
It's designed to capture information we've found to be useful in reviewing pull requests.

If there is other information that would be helpful to include, please don't hesitate to add it!

Please also label your pull request with the relevant tags.
See here for more information and a list of available options:
https://github.com/neurostuff/NiMARE/blob/main/CONTRIBUTING.md#pull-requests
-->

<!-- Please indicate after the # which issue you're closing with this PR.
This is helpful for the maintainers AND will magically close the issue when this
pull request is merged!
https://help.github.com/articles/closing-issues-using-keywords -->
Closes None.

Currently, the `GCLDA` model is slow, mainly when the vocabulary is large (for example for the 3228 terms from Neurosynth). `_update_word_topic_assignments` takes ~73 seconds per iteration and `_update_peak_assignments` takes ~52 seconds per iteration for a total of ~125 seconds of `_update()` per iteration. Compiling and running these functions with `numba` reduce these times to 2, 4, and 6 seconds per iteration respectively.

<!-- Please give a brief overview of what has changed in the PR.
If you're not sure what to write, consider it a note to the maintainers to indicate
what they should be looking for when they review the pull request. -->
Changes proposed in this pull request:

- Convert `_update_word_topic_assignments` and `_update_peak_assignments` to static methods, and add a decorator to compile and run them with `numba` in `nopython` mode.