@tsalo 2022-08-24
With the changes to the two update methods, it seems like they could be converted to functions instead. Is there any benefit to using `@staticmethod` over making them functions?