@JulioAPeraza 2022-08-24
You're right. The two update methods can be converted to functions. I don't see any benefit in using `staticmethod` in this case.