@JulioAPeraza 2022-08-25
Given these limitations (precision issue with sum, and codecov of jitted functions), do you think we should try Cython for these two functions?