@JulioAPeraza 2022-08-26
@adelavega Running times per iteration for 200 topics:

| Function | Branch | Neurosynth (14369 studies, 3228 words) | NeuroQuery (13459 studies, 6145 words) |
| -------- | :--------: | :-----------: | :-----------: |
| `_update_word_topic_assignments` | master | ~96s | ~1342s |
| `_update_word_topic_assignments` | gclda-numba | ~10s | ~149s |
| `_update_peak_assignments` | master | ~112s | ~96s |
| `_update_peak_assignments` | gclda-numba | ~55s | ~50s |

Typically we have 10,000 iterations, so to train GCLDA on Neurosynth with `master` would take ~25 days, which can be reduced to ~7.5 days with `gclda-numba`.