@JulioAPeraza 2022-10-04
The latest changes showed improvement in `_update_peak_assignments`.

Creating chunks (`n_chunks = n_cores`) of the data and running each chunk in separate processors reduced the overhead and the number of `delayed()` calls by the embarrassingly parallel routine. Right now we have 2 nested for loops, one iterates over chunks and it's embarrassingly parallelized, and the other iterates within chunks extending the duration of the task that is submited to Parallel. These changes, besides reducing the overhead, reduce the time that it takes to dispatch so many calls (in our case 500,000) of very fast tasks to workers.

However, this is only scalable to a small number of CPUs. I got small linear scalability for only <= 4 CPUs in a local machine and <= 8 CPUs in an HPC. 

What do you all think? Should we:
1. keep numba support for `_update_word_topic_assignments` and wait for numba to fix the issue with the multinomial generator to add support for `_update_peak_assignments`, or
2. keep both numba support for `_update_word_topic_assignments` and Parallel for `_update_peak_assignments`, with a warning to users that the algorithm is only scalable to a small number of CPUs (4-8)?

I think number 1 is good, since using numba for `_update_word_topic_assignments` already makes GCLDA 10x faster.