@adelavega 2022-10-06
@JulioAPeraza awesome! q: why is it only scalable to a small number of CPUs? I don't quite understand that.