@JulioAPeraza 2022-10-12
I tested the scalability of the function with more CPUs in the HPC and it looks good. I also ran some tests using `threads` but it was slower than the default `loky`.

This PR is ready for review. Thanks!