@adelavega 2022-10-12
Do you know if explicitly setting `parallel=True` would help the numba optimized `_update_peak_assignments`?

https://numba.readthedocs.io/en/stable/user/parallel.html

It looks like the main concern would be the same matrices are being access by the document loop, but it looks like it would be different parts of the arrays?