@JulioAPeraza 2022-10-12
Thanks, @adelavega.

> Do you know if explicitly setting `parallel=True` would help the numba optimized `_update_peak_assignments`?

I tried `parallel=True` before in `_update_word_topic_assignments` and also in `_update_peak_assignments` running on our fork of numba, and I got a `Segmentation Fault` error. I will look into it early tomorrow.